pyproject.toml
# ==========================
# Main project configuration 
# ==========================

[project]
name = "tokenmint"
version = "0.1.0"
description = "A modern Python library for issuing and verifying limited-time, scoped, and replay-protected tokens."
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [
    { name = "shaileshpandit141", email = "shaileshpandit141@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Typing :: Typed",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "authentication",
    "authorization",
    "jwt",
    "tokens",
    "token-auth",
    "time-limited-tokens",
    "ephemeral-tokens",
    "scoped-tokens",
    "secure-tokens",
    "key-rotation",
    "replay-protection",
    "cryptography",
    "EdDSA",
    "Ed25519",
    "Redis",
    "session",
    "magic-link",
    "password-reset",
    "email-verification",
    "api-security",
    "auth",
]
requires-python = ">=3.13"
dependencies = ["cryptography>=45.0.6", "pyjwt>=2.10.1", "redis>=6.4.0"]

# CLI scripts entry point configuration
# -------------------------------------
[project.scripts]
tokenmint = "tokenmint:entry_point"

# ==========================
# Build system configuration
# ==========================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# =================
# Dependency groups
# =================
[dependency-groups]
dev = ["pytest>=8.4.1", "python-dotenv>=1.1.1"]

# =====================
# Optional dependencies
# =====================
[project.optional-dependencies]
dev = ["mypy", "ruff"]

# ===============
# UV tool sources
# ===============
[tool.uv.sources]
tokenmint = { workspace = true }

# ============
# Package data
# ============
[tool.setuptools.package-data]
"tokenmint" = ["py.typed"]

# ============
# Project URLs
# ============
[project.urls]
Homepage = "https://github.com/shaileshpandit141/tokenmint"
Documentation = "https://shaileshpandit141.github.io/tokenmint"
Source = "https://github.com/shaileshpandit141/tokenmint"
Issues = "https://github.com/shaileshpandit141/tokenmint/issues"
License = "https://github.com/shaileshpandit141/tokenmint/blob/main/LICENSE"

# ===================
# Tool Configurations
# ===================

# Disable ruff S101 rule for test casecases
# -----------------------------------------
[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

# ====================
# Pytest configuration
# ====================
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
